[fix/#31] 블록별 시야 조회 DTO 수정

+++ presentation/src/main/java/com/depromeet/presentation/viewfinder/compose/StadiumViewReviewHeader.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.unit.sp
 
 @Composable
 fun StadiumViewReviewHeader(
-    reviewCount: Int,
+    reviewCount: Long,
     modifier: Modifier = Modifier,
     onClickMonthly: () -> Unit
 ) {
